backend/requirements.txt
Django==3.2
django-cors-headers==3.6.0
django-jazzmin==2.6.0
djangorestframework==3.12.2
Pillow==9.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.15.0
requests==2.25.1